# [Console] Improve speed NullOutput