| Q             | A
| ------------- | ---
| Branch?       | 3.4 <!-- see below -->
| Bug fix?      | yes
| New feature?  | no <!-- please update src/**/CHANGELOG.md files -->
| Deprecations? | no <!-- please update UPGRADE-*.md and src/**/CHANGELOG.md files -->
| Tickets       | N/A <!-- prefix each issue number with "Fix #", if any -->
| License       | MIT
| Doc PR        | N/A <!-- required for new features -->

This PR improve NullOutput so that test cases that used it can be run faster.

In my case, the test cases speed is reduced:
* Without blackfire: 4.11s -> 3.08s (reduced 25%)
* With blackfire: 15s -> 7s (reduced 53%)