@nicolas-grekas 2019-12-07
Thank you for the PR @tienvx 
This should target master, as we don't consider perf improvements as bug fixes (unless they are catastrophic, but that's not the case here)